# docs/swagger.yaml
basePath: /
definitions:
...
  doc.User:
    properties:
      active:
        type: boolean
      address_updated:
        type: boolean
      company:
        type: string
      createdat:
        type: string
      designation:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: integer
      image:
        type: string
      is_admin:
        type: boolean
      last_name:
        type: string
      quotas:
        $ref: '#/definitions/models.Quotas'
      reference:
        type: string
      used_demo:
        type: boolean
    type: object
  doc.UserInvite:
    properties:
      company:
        type: string
      createdat:
        type: string
      designation:
        type: string
      email:
        type: string
      email_sent:
        type: boolean
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      remarks:
        type: string
      updatedat:
        type: string
    type: object
...
  models.Quotas:
    properties:
      user_organization:
        type: integer
      user_project:
        type: integer
    type: object
...
host: api.test.01cloud.dev
info:
  contact:
    email: info@01cloud.com
    name: API Support
  description: 01cloud-api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 01cloud-api
  version: 1.0.1
paths:
...
  /user-invite:
    post:
      consumes:
      - application/json
      description: Create User Invite
      parameters:
      - description: Create User Invite
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.UserInvite'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/doc.UserInvite'
      security:
      - ApiKeyAuth: []
      summary: Create User Invite
      tags:
      - UserInvite
  /user-invite-register/{token}:
    get:
      consumes:
      - application/json
      description: Get User Invite
      parameters:
      - description: Get User Invite
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.UserInvite'
      security:
      - ApiKeyAuth: []
      summary: Get User Invite
      tags:
      - UserInvite
    post:
      consumes:
      - application/json
      description: Signup Invite
      parameters:
      - description: Signup Invite
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.User'
      security:
      - ApiKeyAuth: []
      summary: Signup Invite
      tags:
      - UserInvite
  /user-invite-request:
    post:
      consumes:
      - application/json
      description: User Request Demo
      parameters:
      - description: Update Request Demo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.UserInvite'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: User Request Demo
      tags:
      - UserInvite
  /user-invite-validate/{id}:
    post:
      consumes:
      - application/json
      description: User Validate Invite
      parameters:
      - description: Update Validate Invite
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: User Validate Invite
      tags:
      - UserInvite
  /user-invite/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User Invite
      parameters:
      - description: Delete User Invite
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete User Invite
      tags:
      - UserInvite
    put:
      consumes:
      - application/json
      description: Get User Invite
      parameters:
      - description: Get User Invite
        in: path
        name: id
        required: true
        type: integer
      - description: Get User Invite
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.UserInvite'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.UserInvite'
      security:
      - ApiKeyAuth: []
      summary: Get User Invite
      tags:
      - UserInvite
  /user-invites:
    get:
      consumes:
      - application/json
      description: Get User Invites
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/doc.UserInvite'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get User Invites
      tags:
      - UserInvite
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"